Move Euler kernels to separate file
